Move pidRegistry to the registry package
Now that we have a registry package, the pidRegistry can be moved out
of the browser package and into the registry package. This will help
isolate the details.
